+++ CHANGELOG.md
@@ -3,6 +3,13 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.4.1]
+
+### Fixed
+
+- Handle reraising an error from a handler where the caught
+  error(s) is a tuple.
+
 ## [1.4.0]
 
 ### Changed